fix/replace-int-with-uint64 (#35)
* fixes

* removed

+++ Sources/BeaconChain/DataStructures/Operations/Deposit/DepositData.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 struct DepositData {
-    let amount: Int
+    let amount: UInt64
     let timestamp: TimeInterval
     let depositInput: DepositInput
 }
